Added initial implementation for Tetris
Also started developing in linux

+++ GameClones/Tetris/Source/Common/Font.h
@@ -18,7 +18,7 @@ namespace Common {
 			int bearingX;
 			int bearingY;
 
-			int advance;
+			long int advance;
 		};
 
 	private:
